refactoring

+++ src/main/java/com/projectthymeleaf/model/Expense.java
@@ -7,7 +7,7 @@ public class Expense {
 
     Integer id;
     String name;
-    Integer amount;
+    Double amount;
     String cdate;
     String dateEdited;
     TransactionType type;
