+++ src/main/java/com/projectthymeleaf/repository/CalculatorRepository.java
@@ -2,8 +2,8 @@ package com.projectthymeleaf.repository;
 
 public interface CalculatorRepository {
 
-    Integer total();
-    Integer incomeTotal();
-    Integer expenseTotal();
+    Double total();
+    Double incomeTotal();
+    Double expenseTotal();
 
 }
